Add files via upload
This project is a simple text editor built with React and TypeScript that allows users to customize the text appearance using Google Fonts. Users can select a font family, font weight, and toggle italic styling. 

coding platform: VS Code
extension: React, HTML, CSS

How to run: 
   1st TextEditor.tsx  is main file to run command 'npm start' in terminal, 

Future plans of this project are 
if i get some more time i will develop a autosave and reset option for this text that user enters 
and deploy this and make this as a client-server application for making it more easier to access 
and will include many more texts with different styles of fonts ..